part2: ex 2.4, 2.5
