+++ part2/courseinfo/src/App.jsx
@@ -1,60 +1,56 @@
-const Header = ({ course }) => <h1>{course}</h1>
-
-const Total = ({ sum }) => <p><b>total of {sum} exercises</b></p>
-
-const Part = ({ part }) => 
-  <p>
-    {part.name} {part.exercises}
-  </p>
-
-const Content = ({ parts }) => 
-  <>
-    {parts.map(part => <Part key={part.id} part={part} />)}   
-  </>
-
-const Course = ({ course }) => {
-  const { name, parts } = course
-  const sum = parts.reduce((acc, part) => acc + part.exercises, 0)
-
-  return (
-    <>
-      <Header course={name} />
-      <Content parts={parts} />
-      <Total sum={sum} />
-  </>
-  )
-}
+import Course from './components/Course'
 
 const App = () => {
-  const course = {
-    id: 1,
-    name: 'Half Stack application development',
-    parts: [
-      {
-        name: 'Fundamentals of React',
-        exercises: 10,
-        id: 1
-      },
-      {
-        name: 'Using props to pass data',
-        exercises: 7,
-        id: 2
-      },
-      {
-        name: 'State of a component',
-        exercises: 14,
-        id: 3
-      },
-      ,
-      {
-        name: 'Redux',
-        exercises: 11,
-        id: 4
-      }
-    ]
-  }
-
-  return <Course course={course} />
+  const courses = [
+    {
+      name: 'Half Stack application development',
+      id: 1,
+      parts: [
+        {
+          name: 'Fundamentals of React',
+          exercises: 10,
+          id: 1
+        },
+        {
+          name: 'Using props to pass data',
+          exercises: 7,
+          id: 2
+        },
+        {
+          name: 'State of a component',
+          exercises: 14,
+          id: 3
+        },
+        {
+          name: 'Redux',
+          exercises: 11,
+          id: 4
+        }
+      ]
+    }, 
+    {
+      name: 'Node.js',
+      id: 2,
+      parts: [
+        {
+          name: 'Routing',
+          exercises: 3,
+          id: 1
+        },
+        {
+          name: 'Middlewares',
+          exercises: 7,
+          id: 2
+        }
+      ]
+    }
+  ]
+  
+  return (
+    <>
+      {courses.map(course => <Course key={course.id} course={course} />)}
+    </>
+    )
 }
 
 export default App
